participants/itd/source/requirements.txt
aiohttp==3.6.2
alembic==1.4.2
amqp==5.0.2
api.ai==0.1.0
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
blinker==1.4
blueprint==3.4.2
cachetools==4.1.1
celery==5.0.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==2.9.2
decorator==4.4.2
dialogflow==1.0.0
dnspython==1.16.0
email-validator==1.1.1
eventlet==0.29.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
flower==0.9.5
google-api-core==1.21.0
google-auth==1.19.2
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.30.0
humanize==3.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.19.0
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
python-telegram-bot==12.8
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
SpeechRecognition==3.8.1
SQLAlchemy==1.3.17
tornado==6.0.4
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
XlsxWriter==1.3.7
yarl==1.4.2
zope.event==4.5.0
zope.interface==5.2.0
